Unit of work pattern

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 10;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -32,17 +33,21 @@ export class MessagesComponent implements OnInit {
       })
   }
 
-deleteMessage(id: number) {
-  this.messageService.deleteMessage(id).subscribe(() =>{
-    this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-  })
-}
+  deleteMessage(id: number) {
+    this.confirmService.confirm('Confirm delete message', 'This can not be undone. Are you sure to continue?').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
+    })
+  }
 
-pageChanged(event: any) {
-  if (this.pageNumber !== event.page) {
-     this.pageNumber = event.page;
-     this.loadMessages()
-   }
- }
+  pageChanged(event: any) {
+    if (this.pageNumber !== event.page) {
+      this.pageNumber = event.page;
+      this.loadMessages()
+    }
+  }
 
 }
